Dataset of Logistic_Regression Project
Here's a brief description of the columns:

Age: This column represents the age of individuals who may have interacted with digital advertisements. The age is given in years.

Salary: This column represents the salary of the individuals. The salary is provided in a numerical format, but the exact currency is not specified in the sample data.

Status: This column represents the status of whether an individual interacted with a digital advertisement or not. It is a binary column where:

0 indicates that the individual did not interact with the advertisement (Negative class).
1 indicates that the individual interacted with the advertisement (Positive class).